requirements.txt
requests
pynacl
termcolor